combat system and client side player state

+++ src/main/java/de/sealcore/game/items/weapons/WeaponInfo.java
@@ -1,5 +1,5 @@
 package de.sealcore.game.items.weapons;
 
-public record WeaponInfo(WeaponType type, int damage, int range, int magSize)
+public record WeaponInfo(WeaponType type, int damage, double range, int magSize)
 {
 }
